networks config refactor
